refactor: split encoding/encodingwrite.go — plain writes into separate file
Extract plain-encoding write functions into encodingwrite_plain.go; split
matching tests into encodingwrite_plain_test.go.
